app/Scripts/Python/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pymysql>=1.1.0
python-dotenv>=1.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
mysql-connector-python>=8.2.0
googlemaps>=4.10.0
